requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dataclasses-json==0.5.7
decorator==5.1.1
exceptiongroup==1.1.1
fastapi==0.96.0
firebase-admin==6.1.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
httplib2==0.22.0
httptools==0.5.0
idna==3.4
langchain==0.0.189
marshmallow==3.19.0
marshmallow-enum==1.5.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.7
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.2
proto-plus==1.22.2
protobuf==4.23.2
py==1.11.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
pyparsing==3.0.9
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
ratelimiter==1.2.0.post0
regex==2023.6.3
requests==2.28.2
retry==0.9.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.6
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchfiles==0.19.0
weaviate-client==3.19.2
websockets==11.0.3
yarl==1.9.2
